# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Ayah": {
        "description": "One Qur\u02be\u0101n verse, resolved from a surah:ayah reference.\n\n``text_ar`` carries the diacritized recitation text; ``text_plain`` is the\nsame verse with vowel marks stripped, so a reader sees both the pointed and\nthe bare forms that corpus quotations may use. The diacritic-insensitive\nmatching against book content is handled by the search FTS tokenizer, so\neither form locates the same passages.",
        "properties": {
          "ayah": {
            "description": "Ayah number within the surah.",
            "minimum": 1.0,
            "title": "Ayah",
            "type": "integer"
          },
          "surah": {
            "description": "Surah number.",
            "minimum": 1.0,
            "title": "Surah",
            "type": "integer"
          },
          "text_ar": {
            "description": "Verse text with diacritics.",
            "title": "Text Ar",
            "type": "string"
          },
          "text_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English rendering, if available.",
            "title": "Text En"
          },
          "text_plain": {
            "description": "Verse text with diacritics stripped.",
            "title": "Text Plain",
            "type": "string"
          },
          "verse_count": {
            "description": "Total ayat in the surah.",
            "minimum": 1.0,
            "title": "Verse Count",
            "type": "integer"
          }
        },
        "required": ["surah", "ayah", "verse_count", "text_ar", "text_plain"],
        "title": "Ayah",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "The body of every non-2xx response the API declares.",
        "properties": {
          "detail": {
            "description": "Human-readable failure description.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_Ayah_": {
        "properties": {
          "items": {
            "description": "Records in this slice.",
            "items": {
              "$ref": "#/components/schemas/Ayah"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Slice size requested.",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of records skipped before this slice.",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the query, across all pages.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "Page[Ayah]",
        "type": "object"
      },
...
      "Page_PersonEntry_": {
        "properties": {
          "items": {
            "description": "Records in this slice.",
            "items": {
              "$ref": "#/components/schemas/PersonEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Slice size requested.",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of records skipped before this slice.",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the query, across all pages.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "Page[PersonEntry]",
        "type": "object"
      },
...
      "PersonEdge": {
        "description": "A teacher or student relation of a person, linked to a person id when known.",
        "properties": {
          "name": {
            "description": "The related narrator's name as recorded (Arabic).",
            "title": "Name",
            "type": "string"
          },
          "name_latin": {
            "default": "",
            "description": "Latin (ALA-LC-style) reading of the name.",
            "title": "Name Latin",
            "type": "string"
          },
          "other_person_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved person id of the relation, or null when unlinked.",
            "title": "Other Person Id"
          },
          "relation": {
            "description": "'teacher' or 'student'.",
            "title": "Relation",
            "type": "string"
          }
        },
        "required": ["relation", "name"],
        "title": "PersonEdge",
        "type": "object"
      },
      "PersonEntry": {
        "description": "An authoritative narrator identity, cross-checked across its source entries.",
        "properties": {
          "bio": {
            "default": "",
            "description": "Longest available biographical snippet.",
            "title": "Bio",
            "type": "string"
          },
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Birth year, Hijri, when known.",
            "title": "Birth Year"
          },
          "confidence": {
            "default": "medium",
            "description": "Record confidence (high/medium).",
            "title": "Confidence",
            "type": "string"
          },
          "death_conflict": {
            "default": false,
            "description": "Sources disagree on the death year.",
            "title": "Death Conflict",
            "type": "boolean"
          },
          "death_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reconciled Hijri death year.",
            "title": "Death Year"
          },
          "event_count": {
            "description": "Historical events attributed to this person.",
            "minimum": 0.0,
            "title": "Event Count",
            "type": "integer"
          },
          "full_name": {
            "description": "Full name in Arabic.",
            "title": "Full Name",
            "type": "string"
          },
          "generation": {
            "default": "",
            "description": "companion / successor, when derivable.",
            "title": "Generation",
            "type": "string"
          },
          "kunya": {
            "default": "",
            "description": "Teknonym (Abu/Umm ...), if recorded.",
            "title": "Kunya",
            "type": "string"
          },
          "n_sources": {
            "description": "Raw corpus entries merged into this identity.",
            "minimum": 0.0,
            "title": "N Sources",
            "type": "integer"
          },
          "name_latin": {
            "default": "",
            "description": "Approximate ALA-LC transliteration of the name.",
            "title": "Name Latin",
            "type": "string"
          },
          "name_variants": {
            "default": "",
            "description": "Distinct spellings, pipe-separated.",
            "title": "Name Variants",
            "type": "string"
          },
          "nisba": {
            "default": "",
            "description": "Attributive name (tribe/place), if recorded.",
            "title": "Nisba",
            "type": "string"
          },
          "person_id": {
            "description": "Stable person identity id + detail-route key.",
            "minimum": 0.0,
            "title": "Person Id",
            "type": "integer"
          },
          "places": {
            "default": "",
            "description": "Associated places, pipe-separated.",
            "title": "Places",
            "type": "string"
          },
          "reliability": {
            "description": "Per-evaluator reliability grades (evaluator=term).",
            "items": {
              "type": "string"
            },
            "title": "Reliability",
            "type": "array"
          },
          "source_books": {
            "default": "",
            "description": "Source works, pipe-separated.",
            "title": "Source Books",
            "type": "string"
          },
          "student_count": {
            "description": "Distinct recorded students (person_edge).",
            "minimum": 0.0,
            "title": "Student Count",
            "type": "integer"
          },
          "teacher_count": {
            "description": "Distinct recorded teachers (person_edge).",
            "minimum": 0.0,
            "title": "Teacher Count",
            "type": "integer"
          },
          "tradition": {
            "default": "",
            "description": "Sunni / shia / both, when classified.",
            "title": "Tradition",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "person_id",
          "n_sources",
          "teacher_count",
          "student_count",
          "event_count"
        ],
        "title": "PersonEntry",
        "type": "object"
      },
      "PersonEvent": {
        "description": "A historical event attributed to a person.",
        "properties": {
          "event": {
            "default": "",
            "description": "Event name (battle, conquest, ...).",
            "title": "Event",
            "type": "string"
          },
          "event_type": {
            "default": "",
            "description": "Event category (BATTLE, CONQUEST, ...).",
            "title": "Event Type",
            "type": "string"
          },
          "role": {
            "default": "",
            "description": "Marker keyword linking person to event.",
            "title": "Role",
            "type": "string"
          },
          "year_ah": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hijri year of the event, when dated.",
            "title": "Year Ah"
          }
        },
        "title": "PersonEvent",
        "type": "object"
      },
      "PersonGrade": {
        "description": "One reliability grade re-validated against its cited source page.\n\nEach grade names the critic (``evaluator``), the verdict term as it appears in that\ncritic's segment of the narrator's own entry, and the source book + page it was read\nfrom; ``link`` is a relative, URL-independent reader deep-link to exactly that page.\nGrades that could not be located in their cited page are not built, so every row here\nis one a reader can open and confirm.",
        "properties": {
          "book": {
            "default": "",
            "description": "Source work the verdict was read from.",
            "title": "Book",
            "type": "string"
          },
          "evaluator": {
            "default": "",
            "description": "Critic who issued the verdict.",
            "title": "Evaluator",
            "type": "string"
          },
          "link": {
            "default": "",
            "description": "Relative reader deep-link to where it is recorded.",
            "title": "Link",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cited page number in the source work.",
            "title": "Page"
          },
          "term": {
            "default": "",
            "description": "Verdict term, verbatim from the critic's segment.",
            "title": "Term",
            "type": "string"
          }
        },
        "title": "PersonGrade",
        "type": "object"
      },
...
      "Surah": {
        "description": "One full surah: its numbered ayat in recitation order.\n\nThe dedicated Qur\u02be\u0101n reader consumes whole surahs; the longest (al-Baqara,\n286 ayat) is small enough that the full run ships in one response rather\nthan a paged envelope.",
        "properties": {
          "surah": {
            "description": "Surah number.",
            "minimum": 1.0,
            "title": "Surah",
            "type": "integer"
          },
          "verse_count": {
            "description": "Total ayat in the surah.",
            "minimum": 1.0,
            "title": "Verse Count",
            "type": "integer"
          },
          "verses": {
            "description": "The surah's numbered ayat, in order.",
            "items": {
              "$ref": "#/components/schemas/Ayah"
            },
            "title": "Verses",
            "type": "array"
          }
        },
        "required": ["surah", "verse_count", "verses"],
        "title": "Surah",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Knowledge-graph reader for classical Arabic manuscripts.",
    "title": "Shia Online Library",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/person": {
      "get": {
        "description": "Wrap the repo's (slice, total) into a Page[PersonEntry] envelope.",
        "operationId": "_list_person_api_person_get",
        "parameters": [
          {
            "description": "Substring match on name / kunya / nisba / variants.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Substring match on name / kunya / nisba / variants.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Filter by tradition (sunni / shia / both).",
            "in": "query",
            "name": "tradition",
            "required": false,
            "schema": {
              "default": "",
              "description": "Filter by tradition (sunni / shia / both).",
              "title": "Tradition",
              "type": "string"
            }
          },
          {
            "description": "Filter by record confidence (high / medium).",
            "in": "query",
            "name": "confidence",
            "required": false,
            "schema": {
              "default": "",
              "description": "Filter by record confidence (high / medium).",
              "title": "Confidence",
              "type": "string"
            }
          },
          {
            "description": "Only persons with historical events.",
            "in": "query",
            "name": "has_events",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only persons with historical events.",
              "title": "Has Events",
              "type": "boolean"
            }
          },
          {
            "description": "Records per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Records per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Records to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Records to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PersonEntry_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List enriched narrator persons with pagination + filters.",
        "tags": ["person"]
      }
    },
    "/api/person/{person_id}": {
      "get": {
        "description": "Return one enriched person by id, or raise ``ResourceNotFoundError``.",
        "operationId": "get_person_api_person__person_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single enriched person by id.",
        "tags": ["person"]
      }
    },
    "/api/person/{person_id}/edges": {
      "get": {
        "description": "Return a person's teacher/student relations, each split out and transliterated.\n\nA stored edge name may be a comma-joined list of several narrators (Arabic ``\u060c``\nor ASCII ``,``); each is emitted as its own edge so the reader sees one card per\nperson, and each carries a Latin reading from the one shared transliterator\n(the same function that builds ``PersonEntry.name_latin`` at build time).",
        "operationId": "get_person_edges_api_person__person_id__edges_get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonEdge"
                  },
                  "title": "Response Get Person Edges Api Person  Person Id  Edges Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a person's teacher/student relations.",
        "tags": ["person"]
      }
    },
    "/api/person/{person_id}/events": {
      "get": {
        "description": "Return the historical events attributed to a person.",
        "operationId": "get_person_events_api_person__person_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonEvent"
                  },
                  "title": "Response Get Person Events Api Person  Person Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the historical events attributed to a person.",
        "tags": ["person"]
      }
    },
    "/api/person/{person_id}/grades": {
      "get": {
        "description": "Return a person's source-validated reliability grades, each with its reader deep-link.",
        "operationId": "get_person_grades_api_person__person_id__grades_get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonGrade"
                  },
                  "title": "Response Get Person Grades Api Person  Person Id  Grades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a person's source-validated reliability grades with reader links.",
        "tags": ["person"]
      }
    },
    "/api/quran/search": {
      "get": {
        "description": "Wrap the repo's ``(slice, total)`` of matching ayat into a Page[Ayah].",
        "operationId": "_search_verses_api_quran_search_get",
        "parameters": [
          {
            "description": "Arabic term or phrase to find in the Qur\u02be\u0101n.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Arabic term or phrase to find in the Qur\u02be\u0101n.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Records per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Records per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Records to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Records to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Ayah_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Qur\u02be\u0101n verses containing an Arabic term or phrase.",
        "tags": ["quran"]
      }
    },
    "/api/quran/{surah}": {
      "get": {
        "description": "Resolve a surah number to its full run of numbered ayat, in order.\n\nThe prefatory basmala (key ``0``) is skipped, matching the search index:\nit is not a numbered ayah.",
        "operationId": "get_surah_api_quran__surah__get",
        "parameters": [
          {
            "in": "path",
            "name": "surah",
            "required": true,
            "schema": {
              "title": "Surah",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Surah"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve a full surah to its numbered ayat, in order.",
        "tags": ["quran"]
      }
    },
    "/api/quran/{surah}/{ayah}": {
      "get": {
        "description": "Resolve a surah:ayah reference to its verse text (both pointed + bare).",
        "operationId": "get_verse_api_quran__surah___ayah__get",
        "parameters": [
          {
            "in": "path",
            "name": "surah",
            "required": true,
            "schema": {
              "title": "Surah",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "ayah",
            "required": true,
            "schema": {
              "title": "Ayah",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ayah"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve a Qur\u02be\u0101n verse by surah:ayah reference.",
        "tags": ["quran"]
      }
    },
...
  }
}